♻️  [Refactor] : NetworkLoader 파일 나눔
